+++ libappletdecoration/previewshareddecoration.h
@@ -34,7 +34,7 @@ class SharedDecoration : public QObject
     Settings *settings() const;
     void setSettings(Settings *settings);
 
-    KDecoration2::Decoration *decoration() const;
+    KDecoration3::Decoration *decoration() const;
 
     void initDecoration();
 
@@ -51,7 +51,7 @@ class SharedDecoration : public QObject
 
   private:
     QPointer<PreviewBridge> m_bridge;
-    QPointer<KDecoration2::Decoration> m_decoration;
+    QPointer<KDecoration3::Decoration> m_decoration;
     QPointer<Settings> m_settings;
 
     QString m_lastPlugin;